repost by shell.

+++ scripts/build_by_quick.sh
@@ -217,48 +217,71 @@ pre_feeds(){
 }
 pre_feeds
 
-######################## build config (make menuconfig) ########################
+######################## pre build ########################
+# make menuconfig
+do_make_menuconfig(){
+    cd $sdk_path
+    make menuconfig
+}
 default_config(){
     cd $sdk_path
     if [ ! -e .config ]; then
-        diffconfig_file_path="$script_root_path/devices/$device/sdk-diffconfig.info"
+        diffconfig_file_path="$script_root_path/devices/$device/source-diffconfig.info"
         if [ -e $diffconfig_file_path ]; then
             cp -f $diffconfig_file_path .config
             make defconfig
         fi
     fi
 }
-do_menuconfig(){
-    cd $sdk_path
-    make menuconfig
-
-    # @https://p3terx.com/archives/openwrt-compilation-steps-and-commands.html
-    # 查找 dl 目录下文件是否下载正常，小于 1k 的文件，说明下载可能不完整
-    result="find dl -size -1024c -exec ls -l {} \;"
-    if [ -n "$result" ]; then
-        # # 删除 dl 目录下小于 1k 的文件
-        find dl -size -1024c -exec rm -f {} \;
-        make download -j8 V=s
-    fi
-}
 edit_config(){
     cd $sdk_path
+    default_config
     while true; do
         echo -n -e "$INPUT"
         read -p "是否需要修改编译配置 (y/n) ? " yn
         echo
         case $yn in
-            [Yy]* ) do_menuconfig; break;;
+            [Yy]* ) do_make_menuconfig; break;;
             [Nn]* | "" ) break;;
             * ) echo "输入 y 或 n 以确认";;
         esac
     done
 }
-default_config
 edit_config
 
+# make download
+do_make_download(){
+    cd $sdk_path
+    if [ -d dl ]; then
+        # @https://p3terx.com/archives/openwrt-compilation-steps-and-commands.html
+        # 查找 dl 目录下文件是否下载正常，小于 1k 的文件，说明下载可能不完整
+        result="find dl -size -1024c -exec ls -l {} \;"
+        if [ -n "$result" ]; then
+            # # 删除 dl 目录下小于 1k 的文件
+            find dl -size -1024c -exec rm -f {} \;
+            make download -j8 V=s
+        fi
+    else
+        make download -j8 V=s
+    fi
+}
+download_dep(){
+    cd $sdk_path
+    while true; do
+        echo -n -e "$INPUT"
+        read -p "是否需要下载编译依赖 (y/n) ? " yn
+        echo
+        case $yn in
+            [Yy]* ) do_make_download; break;;
+            [Nn]* | "" ) break;;
+            * ) echo "输入 y 或 n 以确认";;
+        esac
+    done
+}
+download_dep
+
 ######################## fix ########################
-# 修复 18.04 动态链接库缺失问题
+# 修复 Ubuntu 18.04 动态链接库缺失问题
 fix_sys(){
     if [ ! -L /lib/ld-linux-x86-64.so.2 ]; then
         sudo ln -s /lib/x86_64-linux-gnu/ld-2.27.so /lib/ld-linux-x86-64.so.2
